requirements.txt
fastapi[all]~=0.110.2
uvicorn
sqlalchemy~=2.0.30
pydantic[email]~=2.7.0
mysqlclient==2.1.0
gssapi~=1.8.3
python-dotenv~=1.0.0
uuid~=1.30
ua-parser
user-agents
pyspnego
logtail-python
requests~=2.31.0
redis~=5.0.4
phonenumbers~=8.13.38
